Create 'lengthToDERLength()' as a module import.
